third_party/WebKit/LayoutTests/bluetooth/server/disconnect/gc-detach.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<body>
<script>
'use strict';
const test_desc = 'Garbage collect then detach frame. We shouldn\'t crash.';

bluetooth_test(() => setUpHealthThermometerDeviceForDiscovery()
  .then(() => createIframe())
  // Connect device, run garbage collection, and detach iframe.
  .then(iframe => new Promise(resolve => {
    let options = {filters: [{services: ['health_thermometer']}]};
    messageIframe(iframe, 'RequestAndConnect', options);

    window.onmessage = messageEvent => {
      assert_equals(messageEvent.data, 'Connected');
      runGarbageCollection().then(() => {
        iframe.remove();
        resolve();
      });
    }
})), test_desc)
</script>
</body>
